pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Simple two-agent chatbot with intent classification and security validation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]
